Change nc to ifconfig to confirm the NIC from the VM to host
Previously, we used "nc" to confirm both the network connectivity and
the successful control switch from the VM to host or the host to VM. It
was inaccurate to measure the time to switch the pNIC control, because
it has counted the network time. Thus, we decided to use "ifconfig" to
check the availability of network interface on the host, when the
control of pNIC has been switched from the VM back to the host
successfully.

modified:   switch_nic_between_host_vm.sh